# server/infra/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MDM Platform API
  version: 1.0.0
  description: |
    Public API surface exposed via the gateway. All endpoints under /api/v1/.
    Auth: Bearer JWT. Two token kinds (claim `knd`):
      - `user`   for admin/operator humans (carries `rol`: viewer|operator|admin|super_admin)
      - `device` for enrolled devices
    Authorization: RBAC enforced per route via a central permission matrix
    (shared/authz). Command issuance is risk-tiered per kind
    (command:issue:basic|privileged|surveillance); user/role management is
    hierarchy-gated. GET /api/v1/auth/roles returns the full matrix.
servers:
  - url: https://mdm.example.com
    description: Production
  - url: http://localhost
    description: Local
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearer: []
paths:
  /api/v1/auth/login:
    post:
      summary: Exchange credentials for an access + refresh token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
                mfa_code: { type: string }
      responses:
        "200":
          description: token pair
...
  /api/v1/auth/users/{id}/role:
    patch:
      summary: "Change a user's role (perm user:manage; no escalation; guards last super_admin)"
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid }}]
  /api/v1/auth/users/{id}/deactivate:
    post:
      summary: "Deactivate a user (perm user:manage; cannot deactivate self or last super_admin)"
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid }}]
...